+++ tests/test_environ.py
@@ -2,12 +2,16 @@
 """Tests the xonsh environment."""
 from __future__ import unicode_literals, print_function
 import os
+import tempfile
+import builtins
 
 import nose
 from nose.tools import (assert_equal, assert_true, assert_not_in,
-                        assert_is_instance, assert_in)
+    assert_is_instance, assert_in, assert_raises)
 
-from xonsh.environ import Env, format_prompt
+from xonsh.environ import Env, format_prompt, load_static_config
+
+from tests.tools import mock_xonsh_env
 
 def test_env_normal():
     env = Env(VAR='wakka')
@@ -73,7 +77,6 @@ def test_HISTCONTROL():
     assert_true('ignoredups' in env['HISTCONTROL'])
 
 def test_swap():
-
     env = Env(VAR='wakka')
     assert_equal(env['VAR'], 'wakka')
 
@@ -100,6 +103,27 @@ def test_swap():
     assert 'VAR1' not in env
     assert 'VAR2' not in env
     assert 'VAR3' not in env
+
+def check_load_static_config(s, exp, loaded):
+    env = {'XONSH_SHOW_TRACEBACK': False}
+    with tempfile.NamedTemporaryFile() as f, mock_xonsh_env(env):
+        f.write(s)
+        f.flush()
+        conf = load_static_config(env, f.name)
+    assert_equal(exp, conf)
+    assert_equal(env['LOADED_CONFIG'], loaded)
+
+def test_load_static_config_works():
+    s = b'{"best": "awash"}'
+    check_load_static_config(s, {'best': 'awash'}, True)
+
+def test_load_static_config_type_fail():
+    s = b'["best", "awash"]'
+    check_load_static_config(s, {}, False)
+
+def test_load_static_config_json_fail():
+    s = b'{"best": "awash"'
+    check_load_static_config(s, {}, False)
 
 
 if __name__ == '__main__':